# Conflicts:
#	build_system/targets/testing/python/requirements.txt

+++ python/tests/mlrl_tests/common/cmd_runner.py
@@ -7,11 +7,14 @@ import subprocess
 from functools import reduce
 from os import environ
 from pathlib import Path
+from typing import List, Optional
 
 import pytest
 
 from .cmd_builder import CmdBuilder
 from .comparison import FileComparison, TextFileComparison
+
+from mlrl.testbed.experiments.state import ExperimentMode
 
 
 class CmdRunner:
@@ -19,15 +22,17 @@ class CmdRunner:
     Allows to run commands that have been configured via a `CmdBuilder`.
     """
 
-    def __format_cmd(self):
-        return reduce(lambda aggr, arg: aggr + (' ' + arg if len(aggr) > 0 else arg), self.args, '')
+    def __format_cmd(self, args: Optional[List[str]] = None) -> str:
+        args = self.args if args is None else args
+        return reduce(lambda aggr, arg: aggr + (' ' + arg if len(aggr) > 0 else arg), args, '')
 
     def __run_cmd(self):
-        out = subprocess.run(self.args, capture_output=True, text=True, check=False)
+        args = self.args
+        out = subprocess.run(args, capture_output=True, text=True, check=False)
         exit_code = out.returncode
 
         if exit_code != 0:
-            pytest.fail('Command "' + self.__format_cmd() + '" terminated with non-zero exit code ' + str(exit_code)
+            pytest.fail('Command "' + self.__format_cmd(args) + '" terminated with non-zero exit code ' + str(exit_code)
                         + '\n\n' + str(out.stderr))
 
         return out
@@ -41,7 +46,8 @@ class CmdRunner:
         expected_files_to_be_deleted = []
 
         for expected_file in expected_output_dir.rglob('*'):
-            if expected_file.is_file() and expected_file != expected_output_dir / 'std.out':
+            if expected_file.is_file() and expected_file != expected_output_dir / 'std.out' and \
+                    not any (parent.name == CmdBuilder.RERUN_DIR.name for parent in expected_file.parents):
                 actual_file = output_dir / expected_file.relative_to(expected_output_dir)
 
                 if self.__should_overwrite_files():
@@ -86,31 +92,48 @@ class CmdRunner:
         self.builder = builder
         self.args = builder.build()
 
-    def run(self, test_name: str):
+    def run(self, test_name: str, wipe_before: bool = True, wipe_after: bool = True, compare_output: bool = True):
         """
         Runs the command.
 
-        :param test_name: The name of the directory that stores the output files produced by the command
+        :param test_name:       The name of the directory that stores the output files produced by the command
+        :param wipe_before:     True, if temporary directory should be deleted before running the command, False
+                                otherwise
+        :param wipe_after:      True, if temporary directory should be deleted after running the command, False
+                                otherwise
+        :param compare_output:  True, if the output of the command should be compared to the expected output, False
+                                otherwise
         """
         builder = self.builder
         output_dir = builder.base_dir
 
         # Delete temporary directories...
-        shutil.rmtree(output_dir, ignore_errors=True)
+        if wipe_before:
+            shutil.rmtree(output_dir, ignore_errors=True)
 
         # Run command...
         out = self.__run_cmd()
 
         if builder.model_save_dir and builder.model_load_dir:
+            shutil.rmtree(output_dir, ignore_errors=True)
             out = self.__run_cmd()
 
         # Check if output of the command is as expected...
-        stdout = [self.__format_cmd()] + str(out.stdout).splitlines()
-        expected_output_dir = builder.expected_output_dir / test_name
-        self.__compare_or_overwrite_files(TextFileComparison(stdout), expected_file=expected_output_dir / 'std.out')
+        if compare_output:
+            stdout = [self.__format_cmd()] + str(out.stdout).splitlines()
+            actual_output_dir = output_dir
+            expected_output_dir = builder.expected_output_dir / test_name
 
-        # Check if all expected files have been created...
-        self.__compare_or_overwrite_output_files(output_dir=output_dir, expected_output_dir=expected_output_dir)
+            if builder.mode in {ExperimentMode.RUN, ExperimentMode.READ} and not builder.show_help:
+                expected_output_dir = expected_output_dir / CmdBuilder.RERUN_DIR
+                actual_output_dir = actual_output_dir / CmdBuilder.RERUN_DIR
+
+            self.__compare_or_overwrite_files(TextFileComparison(stdout), expected_file=expected_output_dir / 'std.out')
+
+            # Check if all expected files have been created...
+            self.__compare_or_overwrite_output_files(output_dir=actual_output_dir,
+                                                     expected_output_dir=expected_output_dir)
 
         # Delete temporary directories...
-        shutil.rmtree(output_dir, ignore_errors=True)
+        if wipe_after:
+            shutil.rmtree(output_dir, ignore_errors=True)
